feat : stat

+++ Assets/0.Work/Agama/Scripts/Core/StatOverride.cs
@@ -13,7 +13,7 @@ namespace Agama.Scripts.Core
     {
         [SerializeField] private StatSO stat;
         [SerializeField] private bool isUseOverride;
-        [SerializeField] private float overrideValue;
+        [SerializeField] private int overrideValue;
 
         public StatOverride(StatSO stat) => this.stat = stat;
